Actually check the size assumptions
We don't want to make assumptions about what the code on the C side
does.  The types in C strictly declare a length, and so we should ensure
that our data is not longer than that.  If it is, the C code is never
even run.

To preserve the API, this causes bottom on size violation for now.