(Dust4go) Initial reimport
Previous commits all had the wrong author identifier; easier to
start over since there weren't that many.

First version that has any bidirectional communication with the basic
structure of both the shaping and the crypto layers in place.  This is
mainly structural; much of the actual crypto is unintegrated, as is any
of the actual Dust modeling so far.

Package names are CamelCase-y for Branding Purposes (apparently?).