Crypting/framing layer revisions: unstick handshake, use KDF
This changes the state machine significantly to allow 32-byte chunks
between the ephemeral public key and the confirmation code so that
we don't get stuck with pending bits untransmittable during model
content encoding.

This also alters the MAC interface in cryptions to allow streaming
and adds a few other functions as needed.