Move crypting framing closer to proto2015
Read and Write of crypting.Session now use datagram rather than
stream semantics (which breaks the non-ensemble TCP use in the
test proxy, but only sort of).  The backpressure relief valve
now uses sequence numbers for synchronizing with the inward-facing
side of the session.  Framing now uses the arbitrary-length zero
padding and the encrypt-then-MAC construction, though in this case
it's currently an encrypt-then-MAC-then-encrypt to avoid slight
weirdness in buffering; that should potentially be redone.

Minor other changes and additions to support this better.  (In
particular, ErrStuck is now gone entirely, as it should have been
some commits ago.)