Begin reworking main suite to handle multiple modes
We're going to need it.  Add link-layer pseudo-addresses using
local pointers as opaque tokens for routing.  Shuffle connection
types around, leaving the (nearly equivalent to before) RawConn
as the "raw" mode; DustProxy now detects the MTU for reblocking
while copying in, but it doesn't do any multi-connection stuff.

The group code is entirely sketches at the moment; it builds,
but none of it has been run.  We probably need to do some
smithing on proc.Link and crypting.Front before finishing this
bit.