Rework identity, mode, and parameters code
Network addresses and nicknames are no longer part of the Dust
package's data model.  EndpointParams and similar are now more
exposed to the client application.  The text file format is
changed somewhat (but this isn't updated in the documentation
yet).  DustProxy invocation is changed accordingly.

Model parameters are now checked for validity by trying to
construct (and discard) a model from them, and so DustTool now
refuses to generate an identity file for a model it doesn't
know about, and takes the privmodels tag at compile time.

The CLI tools now take a mode parameter which must be set,
to pave the way for future "plus" mode using ensembling.
The --incomplete flags are gone since the full "minus" stack
is there even if not absolutely stable yet.