Rearrange packages in preparation for Dust v2 cutoff
The stable version of Dust v2 will have all the potential support
for connection durations, backpressure avoidance, etc. trimmed for
now, because we need to be able to get it into obfs4proxy for this
release cycle.

This also pulls the license information together and adds copyright
headers to all the source files.